feat: create openAPI path filter
